requirements.txt
asgiref==3.8.1
astroid==3.3.9
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==44.0.3
cssselect2==0.7.0
defusedxml==0.7.1
dill==0.4.0
dj-database-url==3.0.1
Django==4.2.10
django-cors-headers==4.7.0
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
fonttools==4.56.0
greenlet==3.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pylint==3.3.7
pyphen==0.17.2
pytest==8.3.4
pytest-django==4.10.0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
SQLAlchemy==2.0.40
sqlparse==0.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.2
typing_extensions==4.13.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.4.0
weasyprint==64.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
gunicorn